requirements.txt
#torch==1.13.0a0+08820cb
#torchvision==0.14.0a0
pathlib==1.0.1
mmcv-full==1.7.0
decord==0.6.0
ftfy==6.1.1
einops==0.6.0
termcolor==2.1.1
timm==0.4.12
regex==2022.10.31
yacs==0.1.8
pandas==1.3.5
hydra-core==1.3.2
wandb==0.16.2
scikit-image==0.21.0